% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets.R
\name{planetnames}
\alias{planetnames}
\title{Returns list of planets}
\usage{
planetnames()
}
\value{
`list` - list of planet names
}
\description{
Returns list of planets
}
\examples{
solar.system::planetnames()
}
