% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets.R
\name{planetnumbers}
\alias{planetnumbers}
\title{Returns number of planets}
\usage{
planetnumbers()
}
\value{
`integer` - number of planets
}
\description{
Returns number of planets
}
\examples{
solar.system::planetnumbers()
}
